Adds initials for Benchmarking AMS
The Benchmark class allows one to specify 'procedure' to test, and various
'measurements' to be made. Both are specified with blocks. When Benchmark#run is
called, it takes initial measurements, then runs the procedure the specified
number of times, recording measurements after each run.

Also, this commit re-runs db/schema.rb, which apparently changes a bit how the
CREATE TABLE statements are generated.